# openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/activities/{legacyId}/travel-guide/generate": {
      "post": {
        "operationId": "TravelGuideController_generate",
        "summary": "Generate travel guide synchronously",
        "parameters": [
          {
            "name": "legacyId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateTravelGuideDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiEnvelopeDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/GenerateTravelGuideResultDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "201": {
            "description": ""
          }
        },
        "tags": [
          "travel-guide"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/activities/{legacyId}/travel-guide/generate-async": {
      "post": {
        "operationId": "TravelGuideController_generateAsync",
        "summary": "Enqueue async travel guide generation job",
        "description": "小程序 callContainer 单次请求 ≤15s，长耗时生成走异步任务 + 轮询。",
        "parameters": [
          {
            "name": "legacyId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateTravelGuideDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiEnvelopeDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/TravelGuideGenerationJobResultDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "201": {
            "description": ""
          }
        },
        "tags": [
          "travel-guide"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/travel-guide/generation-jobs/{jobId}": {
      "get": {
        "operationId": "TravelGuideGlobalController_getGenerationJob",
        "summary": "Poll async travel guide generation job",
        "parameters": [
          {
            "name": "jobId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiEnvelopeDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/TravelGuideGenerationJobResultDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "travel-guide"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/travel-guide/plans/{guideId}": {
      "get": {
        "operationId": "TravelGuideGlobalController_getSavedPlan",
        "summary": "Read saved travel guide plan (public share link)",
        "description": "分享冷启动只读拉取；guideId 即访问凭证。",
        "parameters": [
          {
            "name": "guideId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiEnvelopeDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/TravelGuidePlanReadResultDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "travel-guide"
        ]
      }
    },
    "/api/travel-guide/plans/{guideId}/budget-tier": {
      "patch": {
        "operationId": "TravelGuideGlobalController_selectBudgetTier",
        "summary": "Select budget tier for saved travel guide",
        "parameters": [
          {
            "name": "guideId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SelectTravelGuideBudgetTierDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiEnvelopeDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/TravelGuideBudgetTierResultDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "travel-guide"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/travel-guide/place-suggestions": {
      "get": {
        "operationId": "TravelGuideMapController_placeSuggestions",
        "summary": "Departure city place suggestions",
        "description": "出发地输入提示：高德 inputtips + 本地城市库兜底",
        "parameters": [
          {
            "name": "region",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "keyword",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiEnvelopeDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/PlaceSuggestionsResultDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "travel-guide"
        ]
      }
    },
    "/api/travel-guide/reverse-geocode": {
      "get": {
        "operationId": "TravelGuideMapController_reverseGeocode",
        "summary": "Reverse geocode coordinates to city label",
        "description": "GCJ-02 coordinates → short city/district label for post location metadata.",
        "parameters": [
          {
            "name": "lat",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "lng",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiEnvelopeDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/ReverseGeocodeResultDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "travel-guide"
        ]
      }
    },
...
  },
  "info": {
    "title": "Sync API",
    "description": "Sync App REST API. Responses are wrapped by TransformInterceptor as { code, message, data }.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "ApiEnvelopeDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "number",
            "example": 200
          },
          "message": {
            "type": "string",
            "example": "success"
          },
          "data": {
            "type": "object",
            "description": "Response payload"
          }
        },
        "required": [
          "code",
          "message",
          "data"
        ]
      },
...
      "TravelGuidePlanDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "summary": {
            "type": "string"
          }
        },
        "required": [
          "title"
        ]
      },
      "GenerateTravelGuideResultDto": {
        "type": "object",
        "properties": {
          "plan": {
            "$ref": "#/components/schemas/TravelGuidePlanDto"
          },
          "guideId": {
            "type": "string"
          }
        },
        "required": [
          "plan"
        ]
      },
      "GenerateTravelGuideDto": {
        "type": "object",
        "properties": {
          "guideId": {
            "type": "string",
            "description": "客户端生成的攻略 ID，用于分享冷启动与服务端只读拉取",
            "minLength": 8,
            "maxLength": 128
          },
          "departure": {
            "type": "string",
            "minLength": 1
          },
          "departureCity": {
            "type": "string",
            "description": "POI 补全返回的城市（如「上海市」），地理编码时优先于活动举办城市"
          },
          "headcount": {
            "type": "number",
            "minimum": 1,
            "maximum": 10
          },
          "budgetTier": {
            "type": "string",
            "enum": [
              "economy",
              "standard",
              "comfort"
            ]
          },
          "selfDrive": {
            "type": "boolean"
          },
          "accommodationNights": {
            "type": "number",
            "minimum": 0,
            "maximum": 7
          }
        },
        "required": [
          "departure",
          "headcount"
        ]
      },
      "TravelGuideGenerationJobResultDto": {
        "type": "object",
        "properties": {
          "jobId": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "running",
              "completed",
              "failed"
            ]
          },
          "plan": {
            "$ref": "#/components/schemas/TravelGuidePlanDto"
          },
          "errorMessage": {
            "type": "string"
          }
        },
        "required": [
          "jobId",
          "status"
        ]
      },
      "TravelGuidePlanReadResultDto": {
        "type": "object",
        "properties": {
          "guideId": {
            "type": "string"
          },
          "activityLegacyId": {
            "type": "number"
          },
          "plan": {
            "$ref": "#/components/schemas/TravelGuidePlanDto"
          },
          "createdAt": {
            "type": "string"
          }
        },
        "required": [
          "guideId",
          "activityLegacyId",
          "plan",
          "createdAt"
        ]
      },
      "TravelGuideBudgetTierResultDto": {
        "type": "object",
        "properties": {
          "guideId": {
            "type": "string"
          },
          "budgetTier": {
            "type": "string"
          }
        },
        "required": [
          "guideId",
          "budgetTier"
        ]
      },
      "SelectTravelGuideBudgetTierDto": {
        "type": "object",
        "properties": {
          "budgetTier": {
            "type": "string",
            "enum": [
              "economy",
              "standard",
              "comfort"
            ]
          }
        },
        "required": [
          "budgetTier"
        ]
      },
      "PlaceSuggestionItemDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "city": {
            "type": "string"
          }
        },
        "required": [
          "title",
          "address"
        ]
      },
      "PlaceSuggestionsResultDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PlaceSuggestionItemDto"
            }
          }
        },
        "required": [
          "data"
        ]
      },
      "ReverseGeocodeResultDto": {
        "type": "object",
        "properties": {
          "label": {
            "type": "string"
          }
        }
      },
...
    }
  }
}